here you go, an automatic scan-finding mail script
We've recently started tracking issues with severity treshhold higher than Medium in another call system for better internal processing and reporting.
I used to copy/paste all the stuff over; but after an internal scan where more than 400 findings where above the treshhold I figured it can be done smarter.
Enter mailSeccubusFindings, a python script that will, using the seccubus API:
For one workspace, go through all scans and get all results ofthose scans where status =1 (NEW). If the result severity matches one of the defined strings and it has not been previously emailed, it will email the finding, and then update the finding with the string (to prevent re-mailing of multiple issues)

https://github.com/Ar0xA/Seccubus/blob/master/scripts/mailSeccubusFindings.py
I do not have time to make this a feature in Seccubus at the moment, so this will have to do :)