Fix NBE report parsing with OpenVAS.
With current debian versions of OpenVAS retrieving the report breaks. Original code was very fragile, in fact.
Replaced the search for the report content with a regular expression that should not break so easily.

Note: The OpenVAS XML output does not validate as proper XML.
That's the root of this problem - since it can't be validated the script resorts to manual text parsing, which naturally breaks when the output format changes even slightly.
Note2: This code does not check if we actually got a result. That still needs to be added, but if that is going to be done it should be done with *everything* in this script, not just the bit that gets the report from OpenVAS's response.